+++ kt/plugins/godot-gradle-plugin/src/main/kotlin/godot/gradle/tasks/packageBootstrapJarTask.kt
@@ -8,7 +8,7 @@ import org.gradle.api.tasks.TaskProvider
 fun Project.packageBootstrapJarTask(): TaskProvider<out Task> {
     return tasks.register("packageBootstrapJar", ShadowJar::class.java) {
         with(it) {
-            group = "godot-kotlin-jvm"
+            group = "godot-jvm"
             description = "Creates a fat jar containing everything needed to load and run the main.jar"
 
             archiveBaseName.set("godot-bootstrap")