+++ src/gd_kotlin.cpp
@@ -417,7 +417,7 @@ Object* GDKotlin::get_callable_middleman() const {
 }
 
 void GDKotlin::display_initialization_error_hint(String cause, String hint) {
-    String warning {"Godot Kotlin/JVM module couldn't be fully initialized.\n"
+    String warning {"Godot-JVM module couldn't be fully initialized.\n"
                     "Java and Kotlin scripts will still appear in the editor but won't be functional.\n"
                     "The cause was:\n"};
     String pre_hint {"\nOne possible solution is:\n"};